refactor report api & type ints
