+++ cg/cli/upload/delivery_report.py
@@ -6,7 +6,7 @@ import click
 
 from cg.cli.generate.report.helpers import resolve_report_case, resolve_report_api
 from cg.cli.generate.report.options import ARGUMENT_CASE_ID
-from cg.meta.report.api import ReportAPI
+from cg.meta.report.report_api import ReportAPI
 
 LOG = logging.getLogger(__name__)
